fix: moved manage profile and password to profile containers
