+++ src/modules/Profile/containers/ManagePassword/index.module.scss
